Compiling

+++ cu_slicer_v2/util/types.h
@@ -9,8 +9,8 @@
 // Flattenned array takes .5 gb for all partitions
 // Reorder - map 
 // Nodes [753565, 226407, 36803, 4096] Edges [1743108, 308245, 33447]
-
-typedef short PARTITIONIDX ;
-
-typedef  int NDTYPE;
+// typedef short PARTITIONIDX ;
+typedef int PARTITIONIDX;
+typedef long NDTYPE;
+// typedef  int NDTYPE;
 // All graphs have less than 2 billion edges
